Check the user's permissions when running the holdings report
On the frontend, we don't want to show the user repositories that they
don't have access to view.  On the backend, the report should check that
the job owner has the right permissions to run the report.

This required passing the job as a parameter to reports, as we need this
to determine the owner of the report.