Allow Solr's "mm" parameter to be used
Solr's `mm` ("must match") parameter lets you define rules for how
many query terms must match for a given document to be returned.  It
lets you say things like "if there are only two search terms, require
both to match; otherwise, require 75% of terms to match".  You can set
this on a per-query basis, or on a global basis.

The way that the PUI constructs queries meant that `mm` wasn't working
as expected.  There were several places that were adding extra clauses
to the search query that could more properly be expressed as search
filters (i.e. the purpose of these extra clauses was to restrict the
search to a subset, not to influence relevance ranking).

This patch reworks these places to use filters instead of adding to
the search query.  It also corrects some of the places where the
filters worked by accident due to the permissiveness of the default OR
queries.